refactor(spec): extract dataChecks layer walk; per-layer style/color evidence (#844)
Move geom/stat rules and channel collection into validate-data-checks-layer.ts
so dataChecks is a thin orchestrator. Style and color checkers now use
evidenceForUse like position, so multi-table same-name fields keep independent
types and no longer last-wins-hide scale-type-mismatch diagnostics.
