+++ packages/spec/src/validate-data-checks.ts
@@ -1,30 +1,33 @@
 /**
- * Tier-2 data-aware walk — field existence, all-null columns, {stat} columns,
- * geom/stat type rules, channel collection, then scale/type compatibility.
+ * Tier-2 data-aware orchestrator — resolve evidence, walk layers, then scale/type
+ * compatibility (style → position → color diagnostic order).
  *
  * Implementation:
+ *  - validate-data-checks-layer.ts — geom/stat rules + field walk + channel collection
  *  - validate-data-checks-temporal.ts — shared temporalDecisionForField memoization
  *  - validate-data-checks-position.ts — pre-evidence temporal axis config + x/y scale types
  *  - validate-data-checks-color.ts — color/fill scale types + manual domain/range
  *  - validate-data-checks-style.ts — shape/linetype + size/linewidth/alpha scale types
  *
+ * Style and color scale checks take the same per-use `evidenceForUse` path as
+ * position (#609 / #844): multi-table layers that share a field name keep their
+ * own type evidence. The last-wins union `fields` map remains a fallback only.
+ *
  * Consumes a FieldEvidenceMap built by validate-data-evidence.ts (or builds one
  * when not pre-resolved).
  */
 import type { SpecError } from "./errors.js";
-import { didYouMean } from "./errors.js";
 import type { Aes, ChannelName } from "./schema.js";
-import { CHANNELS, GEOM_DEFAULTS } from "./schema.js";
-import type { ProfileFieldType, ValidateLimits, ValidateOptions } from "./validate-data.js";
+import type { ValidateLimits, ValidateOptions } from "./validate-data.js";
 import type { TemporalDecision } from "./temporal-column.js";
 import {
-  effectiveChannel,
   resolveFieldEvidence,
   resolveLayerFieldEvidence,
   type FieldEvidenceMap,
   type ResolveFieldEvidenceResult,
 } from "./validate-data-evidence.js";
 import { checkColorScaleDataCompatibility } from "./validate-data-checks-color.js";
+import { collectLayerDataChecks, STAT_COLUMNS } from "./validate-data-checks-layer.js";
 import {
   checkPositionScaleDataCompatibility,
   scaleRequestsTime,
@@ -34,35 +37,8 @@ import {
   checkFiniteStyleScaleDataCompatibility,
   checkNumericStyleScaleDataCompatibility,
 } from "./validate-data-checks-style.js";
-import type { ChannelFieldUse } from "./validate-data-checks-temporal.js";
 
-/** Columns generated by each stat ({ stat } channel resolution contract).
- *  Documented per stat in its @ggsvelte/core module header. */
-export const STAT_COLUMNS: Record<string, readonly string[]> = {
-  identity: [],
-  count: ["count"],
-  bin: ["count", "density", "ncount", "ndensity"],
-  bin_hex: ["count", "density", "ncount", "ndensity"],
-  bin_2d: ["count", "density", "ncount", "ndensity"],
-  density: ["density", "count", "scaled", "ndensity"],
-  ydensity: ["density", "count", "scaled", "violinwidth", "y"],
-  bindot: ["stackpos"],
-  smooth: ["y", "ymin", "ymax", "se"],
-  boxplot: ["ymin", "lower", "middle", "upper", "ymax"],
-  summary: ["y", "ymin", "ymax"],
-  /** after_stat n / prop for geom_count; not published on y (y is position input). */
-  sum: ["n", "prop"],
-  function: ["y"],
-  /** ggplot2 after_stat(ecdf); y channel defaults to this column. */
-  ecdf: ["ecdf"],
-  summary_bin: ["y", "ymin", "ymax"],
-  contour: ["level"],
-  quantile: ["y"],
-  density_2d: ["level", "density"],
-  density_2d_filled: ["level", "density"],
-  qq: ["sample", "theoretical"],
-  qq_line: ["sample", "theoretical"],
-};
+export { STAT_COLUMNS };
 
 function isRecord(v: unknown): v is Record<string, unknown> {
   return typeof v === "object" && v !== null && !Array.isArray(v);
@@ -112,269 +88,20 @@ export function dataChecks(
     return axis !== null && scaleRequestsTime(scales, axis);
   };
 
-  const axisFields: Record<"x" | "y", ChannelFieldUse[]> = { x: [], y: [] };
-  const colorFields: Record<"color" | "fill", ChannelFieldUse[]> = { color: [], fill: [] };
-  const colorScaledConstants: Record<"color" | "fill", unknown[]> = { color: [], fill: [] };
-  const finiteStyleFields: Record<"shape" | "linetype", ChannelFieldUse[]> = {
-    shape: [],
-    linetype: [],
-  };
-  const numericStyleFields: Record<"size" | "linewidth" | "alpha", ChannelFieldUse[]> = {
-    size: [],
-    linewidth: [],
-    alpha: [],
-  };
-  const numericStyleScaledConstants: Record<"size" | "linewidth" | "alpha", unknown[]> = {
-    size: [],
-    linewidth: [],
-    alpha: [],
-  };
-
-  for (let i = 0; i < layers.length; i++) {
-    const layer = layers[i];
-    if (!isRecord(layer)) continue;
-    const geom = typeof layer["geom"] === "string" ? layer["geom"] : "";
-    const layerAes = isRecord(layer["aes"]) ? (layer["aes"] as Aes) : undefined;
-    const defaultStat = GEOM_DEFAULTS[geom as keyof typeof GEOM_DEFAULTS]?.stat ?? "identity";
-    const stat = typeof layer["stat"] === "string" ? layer["stat"] : defaultStat;
-
-    const fields: FieldEvidenceMap | null =
-      layerResolved.status === "ok" ? (layerResolved.layers[i] ?? plotFields) : plotFields;
-    if (fields === null) continue; // runtime-only named data: skip field checks
-    const available = [...fields.keys()];
-
-    // --- geom/stat field-type rules (M2 statistical layer) -----------------
-    const fieldTypeOf = (channel: ChannelName): [string, ProfileFieldType] | null => {
-      const m = effectiveChannel(plotAes, layerAes, channel);
-      if (m === undefined || !("field" in m)) return null;
-      const t = fields.get(m.field)?.type;
-      if (t === undefined || t === null) return null;
-      // Geom/stat contracts see position values after configured temporal parsing.
-      // Parse failures and configuration errors are owned by scale validation below.
-      return [m.field, scaleRequestsTimeForChannel(channel) ? "temporal" : t];
-    };
-    const typeError = (channel: ChannelName, message: string, fixDesc: string) => {
-      errors.push({
-        code: "channel-type-mismatch",
-        path: `/layers/${i}/aes/${channel}`,
-        message,
-        fix: { description: fixDesc },
-      });
-    };
-    if (geom === "smooth") {
-      for (const channel of ["x", "y"] as const) {
-        const info = fieldTypeOf(channel);
-        if (info !== null && (info[1] === "nominal" || info[1] === "ordinal")) {
-          typeError(
-            channel,
-            `The smooth stat needs quantitative x and y, but field "${info[0]}" (${channel}) is ${info[1]}.`,
-            "Map the channel to a numeric field, or use a boxplot for discrete x.",
-          );
-        }
-      }
-    }
-    if (geom === "boxplot") {
-      const x = fieldTypeOf("x");
-      if (x !== null && (x[1] === "quantitative" || x[1] === "temporal")) {
-        typeError(
-          "x",
-          `The boxplot geom needs a DISCRETE x this milestone, but field "${x[0]}" is ${x[1]}.`,
-          "Map x to a categorical field (strings), or bin the values into labeled categories first.",
-        );
-      }
-      const y = fieldTypeOf("y");
-      if (y !== null && (y[1] === "nominal" || y[1] === "ordinal")) {
-        typeError(
-          "y",
-          `The boxplot stat needs a quantitative y, but field "${y[0]}" is ${y[1]}.`,
-          "Map y to a numeric field.",
-        );
-      }
-    }
-    if (
-      geom === "histogram" ||
-      geom === "freqpoly" ||
-      geom === "density" ||
-      (geom === "bar" && stat === "bin") ||
-      (geom === "line" && stat === "bin")
-    ) {
-      const x = fieldTypeOf("x");
-      if (x !== null && (x[1] === "nominal" || x[1] === "ordinal")) {
-        typeError(
-          "x",
-          `The ${geom === "density" ? "density" : "bin"} stat needs a continuous x, but field "${x[0]}" is ${x[1]}.`,
-          'Use geom "bar" (the count stat) to count categories instead.',
-        );
-      }
-    }
-    if (
-      geom === "errorbar" ||
-      geom === "linerange" ||
-      geom === "pointrange" ||
-      geom === "crossbar"
-    ) {
-      for (const channel of ["ymin", "ymax"] as const) {
-        const info = fieldTypeOf(channel);
-        if (
-          info !== null &&
-          (info[1] === "nominal" || info[1] === "ordinal") &&
-          !scaleRequestsTime(scales, "y")
-        ) {
-          typeError(
-            channel,
-            `The ${geom} geom needs quantitative bounds, but field "${info[0]}" (${channel}) is ${info[1]}.`,
-            "Map the channel to a numeric field.",
-          );
-        }
-      }
-    }
-    if (geom === "rect") {
-      for (const channel of ["xmin", "xmax"] as const) {
-        const info = fieldTypeOf(channel);
-        if (
-          info !== null &&
-          (info[1] === "nominal" || info[1] === "ordinal") &&
-          !scaleRequestsTime(scales, "x")
-        ) {
-          typeError(
-            channel,
-            `The rect geom needs quantitative edges, but field "${info[0]}" (${channel}) is ${info[1]}.`,
-            "Map the channel to a numeric field.",
-          );
-        }
-      }
-      for (const channel of ["ymin", "ymax"] as const) {
-        const info = fieldTypeOf(channel);
-        if (
-          info !== null &&
-          (info[1] === "nominal" || info[1] === "ordinal") &&
-          !scaleRequestsTime(scales, "y")
-        ) {
-          typeError(
-            channel,
-            `The rect geom needs quantitative edges, but field "${info[0]}" (${channel}) is ${info[1]}.`,
-            "Map the channel to a numeric field.",
-          );
-        }
-      }
-    }
-    if (geom === "raster") {
-      for (const channel of ["x", "y"] as const) {
-        const info = fieldTypeOf(channel);
-        if (info !== null && (info[1] === "nominal" || info[1] === "ordinal")) {
-          typeError(
-            channel,
-            `The raster geom needs continuous ${channel}, but field "${info[0]}" is ${info[1]}.`,
-            'Use geom "tile" for discrete axes.',
-          );
-        }
-      }
-    }
-    if (geom === "tile") {
-      for (const channel of ["width", "height"] as const) {
-        const info = fieldTypeOf(channel);
-        if (info !== null && info[1] !== "quantitative") {
-          typeError(
-            channel,
-            `The tile geom needs quantitative ${channel}, but field "${info[0]}" is ${info[1]}.`,
-            `Map ${channel} to a positive numeric field, or use params.${channel}.`,
-          );
-        }
-      }
-    }
-
-    for (const channel of CHANNELS) {
-      const mapped = effectiveChannel(plotAes, layerAes, channel);
-      if (mapped === undefined) continue;
-      const path = `/layers/${i}/aes/${channel}`;
-
-      if ("stat" in mapped) {
-        const generated = STAT_COLUMNS[stat] ?? [];
-        if (!generated.includes(mapped.stat)) {
-          errors.push({
-            code: "unknown-stat-column",
-            path,
-            message:
-              generated.length === 0
-                ? `Channel "${channel}" maps stat column "${mapped.stat}", but this layer's stat ("${stat}") generates no columns.`
-                : `Channel "${channel}" maps stat column "${mapped.stat}", but this layer's stat ("${stat}") generates: ${generated.join(", ")}.`,
-            ...(generated.length > 0 && { allowed: [...generated] }),
-          });
-        }
-        continue;
-      }
-      if ("value" in mapped) {
-        // Runtime trains manual/ordinal color domains on scaled constants too.
-        if (
-          (channel === "color" || channel === "fill") &&
-          mapped.scale === true &&
-          mapped.value !== null
-        ) {
-          colorScaledConstants[channel].push(mapped.value);
-        }
-        // The runtime trains numeric style scales on scaled constants as well, so
-        // a scaled string constant on a sequential/binned size/linewidth/alpha
-        // scale must face the same scale-family check as a mapped field below.
-        if (
-          (channel === "size" || channel === "linewidth" || channel === "alpha") &&
-          mapped.scale === true &&
-          mapped.value !== null
-        ) {
-          numericStyleScaledConstants[channel].push(mapped.value);
-        }
-        continue;
-      }
-      if (!("field" in mapped)) continue;
-
-      const info = fields.get(mapped.field);
-      if (info === undefined) {
-        const suggestion = didYouMean(mapped.field, available);
-        errors.push({
-          code: "unknown-field",
-          path,
-          message:
-            `Unknown field "${mapped.field}" (available: ${available.join(", ") || "none"}).` +
-            (suggestion === undefined ? "" : ` Did you mean "${suggestion}"?`),
-          allowed: available,
-          ...(suggestion !== undefined && {
-            fix: {
-              description: `Map "${channel}" to "${suggestion}".`,
-              example: { field: suggestion },
-            },
-          }),
-        });
-        continue;
-      }
-      if (info.allNull) {
-        errors.push({
-          code: "all-null-column",
-          path,
-          message: `Field "${mapped.field}" contains only null values; the "${channel}" channel cannot be drawn from it.`,
-        });
-        continue;
-      }
-      if (channel === "x" || channel === "xmin" || channel === "xmax" || channel === "xend") {
-        axisFields.x.push({ field: mapped.field, path });
-      }
-      if (channel === "y" || channel === "ymin" || channel === "ymax" || channel === "yend") {
-        axisFields.y.push({ field: mapped.field, path });
-      }
-      if (channel === "color" || channel === "fill") {
-        colorFields[channel].push({ field: mapped.field, path });
-      }
-      if (channel === "shape" || channel === "linetype") {
-        finiteStyleFields[channel].push({ field: mapped.field, path });
-      }
-      if (channel === "size" || channel === "linewidth" || channel === "alpha") {
-        numericStyleFields[channel].push({ field: mapped.field, path });
-      }
-    }
-  }
+  const layerMaps = layerResolved.status === "ok" ? layerResolved.layers : null;
+  const walk = collectLayerDataChecks({
+    layers,
+    plotAes,
+    plotFields,
+    layerMaps,
+    scaleRequestsTime: scaleRequestsTimeForChannel,
+  });
+  errors.push(...walk.errors);
 
   // --- scale/type compatibility (order preserved for diagnostics) ------------
-  // Union layer evidence as a fallback for non-layer-scoped checks (e.g. color
-  // manual ranges). Position scale checks prefer per-layer evidence via
-  // evidenceForUse so same field names on different tables stay independent (#609).
+  // Union layer evidence as a fallback for non-layer-scoped lookups. Position,
+  // style, and color scale checks prefer per-layer evidence via evidenceForUse
+  // so same field names on different tables stay independent (#609 / #844).
   const fields: FieldEvidenceMap = new Map(plotFields ?? undefined);
   if (layerResolved.status === "ok") {
     for (const layerMap of layerResolved.layers) {
@@ -385,7 +112,6 @@ export function dataChecks(
   if (fields.size === 0 && resolved.status === "ok") {
     for (const [name, entry] of resolved.fields) fields.set(name, entry);
   }
-  const layerMaps = layerResolved.status === "ok" ? layerResolved.layers : null;
   const evidenceForUse = (use: { field: string; path: string }) => {
     const match = /^\/layers\/(\d+)\//.exec(use.path);
     if (match !== null && layerMaps !== null) {
@@ -403,33 +129,35 @@ export function dataChecks(
     return fields.get(use.field);
   };
   // Diagnostic order: finite style → numeric style → position → color.
-  // Style checkers use the last-wins union `fields` map (not evidenceForUse).
   errors.push(
     ...checkFiniteStyleScaleDataCompatibility({
       scales,
       fields,
-      finiteStyleFields,
+      evidenceForUse,
+      finiteStyleFields: walk.finiteStyleFields,
     }),
     ...checkNumericStyleScaleDataCompatibility({
       scales,
       fields,
-      numericStyleFields,
-      numericStyleScaledConstants,
+      evidenceForUse,
+      numericStyleFields: walk.numericStyleFields,
+      numericStyleScaledConstants: walk.numericStyleScaledConstants,
       temporalDecisionCache,
     }),
     ...checkPositionScaleDataCompatibility({
       scales,
       fields,
       evidenceForUse,
-      axisFields,
+      axisFields: walk.axisFields,
       invalidTemporalAxes,
       temporalDecisionCache,
     }),
     ...checkColorScaleDataCompatibility({
       scales,
       fields,
-      colorFields,
-      colorScaledConstants,
+      evidenceForUse,
+      colorFields: walk.colorFields,
+      colorScaledConstants: walk.colorScaledConstants,
       temporalDecisionCache,
     }),
   );
